requirements.txt
pandas>=2.0
openpyxl>=3.1
lxml>=4.9
PyYAML>=6.0
requests>=2.28
